♻️ separate the register form in 2 steps
